requirements.txt
kurigram>=2.2.17
pytgcrypto>=1.2.11
py-tgcalls~=2.2.11
python-dotenv~=1.2.1
